Create rejseplanen.dk.dmfr.json (#593)
* Create rejseplanen.dk.dmfr.json

I used the Samtrafiken file as a template, as I couldn't find the example file. For now, I've just added the data provider as the operator. Should I add all operators/agencies instead? They are included in the GTFS file, so using those instead would make it easier to create and maintain them. Thanks!

* Update rejseplanen.dk.dmfr.json

Removed the operator per recommendation from drewda, as operators are included in the GTFS.

* `transitland dmfr format --save`

Co-authored-by: Drew Dara-Abrams <drew@interline.io>