Roseville Transit (#1771)
* Updated Roseville Transit, California, USA

Background: Prior to August 2025, Roseville Transit partnered with
Sacramento Regional Transit (SacRT) to host a static GTFS feed, and
with GMV for a GTFS-RT feed. As of October 2025, Roseville Transit
changed data providers to Tripshot, which now hosts the replacement
static and real-time feeds. The GMV feed is no longer operational,
and the SacRT static feed is seemingly out-of-date.

Changes:
 - f-9qcs-rosevilletransit moved from sacrt.com.dmfs.json to
   tripshot.com.dmfr.json, with new feed URL and additional operator
   data
 - f-rosevilletransit~rt deprecated, and replaced with
   f-9qcs-rosevilletransit~rt, to match the static GTFS feed's Onestop ID,
   and using the new real-time feed URLs

References:
https://reports.dds.dot.ca.gov/gtfs_schedule/2025/09/271/index.html
https://cityofroseville.tripshot.com/
https://www.roseville.ca.us/news/what_s_happening_in_roseville/plan_track_ride_with_roseville_transits_new_app

* adjustments

---------

Co-authored-by: 3DPGG <3dpgg@protonmail.com>